shared/requirements.txt
requests>=2.31.0
pandas>=2.0.0
google-cloud-bigquery>=3.11.0
google-cloud-bigquery-storage>=2.22.0
db-dtypes>=1.1.1
gspread>=5.10.0
gspread-dataframe>=3.3.0
google-auth>=2.22.0
tenacity>=8.2.0
